feat: separate routes and handlers & auth functionality & JWT implementation
